refactor(native): move the native tool implementations into their own module
The last region moves to `api_process_adapter/native_tools.rs`: skill input
types and reads, the registered-tool recorder and dispatch, todo write, shell
background/output/kill, `execute_tool_call_impl`, workspace mutation publishing,
the code-intelligence tools, and remember/recall/search. It also takes
`parse_optional_non_negative_integer_arg` and `non_negative_integer`, the
tool-argument parsers that design.md's census had grouped with the interactive
tools.

`mod.rs` is left with what tasks.md 3.9 asked for — the adapter struct, its
three impls, the constants and type aliases — plus the three `GenerationProcessEvent`
constructors (`failed_non_retryable`, `failed_configuration`, `failed_retryable`)
that tasks.md assigned to no module and that four other modules call.

A `[ARCH-NATIVE-006]` entry records this module at 1,478 lines.
