+++ src-tauri/src/contexts/agent_runtime/infrastructure/api_process_adapter/execution.rs
@@ -19,14 +19,17 @@ use super::invocation::{
     begin_api_invocation, estimated_input_characters, finish_api_invocation,
     record_context_snapshot, wire_format_for,
 };
+use super::native_tools::{
+    execute_registered_native_tool, execute_tool_call_with_runtime_ports, is_image_read_request,
+    log_image_attachment, resolve_tool_image,
+};
 use super::prompt::{
     resolve_personalization_settings, resolve_system_prompt_with_settings,
     resolve_tool_catalog_with_code_intelligence,
 };
 use super::{
-    execute_registered_native_tool, execute_tool_call_with_runtime_ports, failed_configuration,
-    failed_non_retryable, failed_retryable, is_image_read_request, log_image_attachment,
-    resolve_tool_image, ExecutedToolCall, PendingApprovals, HISTORY_LIMIT, MAX_TOOL_ROUND_TRIPS,
+    failed_configuration, failed_non_retryable, failed_retryable, ExecutedToolCall,
+    PendingApprovals, HISTORY_LIMIT, MAX_TOOL_ROUND_TRIPS,
 };
 use crate::contexts::agent_runtime::application::{
     delegate_utility_skill_tool_definition, AgentClockPort, AgentCodeIntelligenceContext,
